// sanity.types.ts
export type Project = {
  _id: string;
  _type: "project";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title: string;
  slug: Slug;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  description?: string;
  challenges?: Array<{
    title?: string;
    description?: string;
    _type: "challenge";
    _key: string;
  }>;
  shortDescription?: string;
  learning?: string;
  stack?: Array<{
    name?: string;
    type?: string;
    _type: "stack-item";
    _key: string;
  }>;
  links?: Array<{
    name?: string;
    url?: string;
    _type: "link-item";
    _key: string;
  }>;
  tags?: Array<string>;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top: number;
  bottom: number;
  left: number;
  right: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x: number;
  y: number;
  height: number;
  width: number;
};

export type Slug = {
  _type: "slug";
  current: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;